a ton of things
- added online ranking system manager which works with SilentWolf plugin (it's not included in commit because it basically has no license and I'm not sure would I be able to put it here)
- added shuffle and playlist clear buttons to playlist mode menu
- reworked text input window and accept window source code to make it easier to use
- reworked square checking and timeline erasement code to fix some bugs and make whole system more optimized and accurate

+++ menu/scripts/screens/subscreens/profile_edit.gd
@@ -52,7 +52,9 @@ func _display_profile_list() -> void:
 	if profiles.is_empty() : 
 		$V/Text.text = "NO PROFILES FOUND"
 		$V/Delete._disabled(true)
-	
+	elif profiles.size() < 2:
+		$V/Delete._disabled(true)
+
 	var tween : Tween = create_tween()
 	tween.tween_interval(0.75)
 	var count : int = 0
@@ -90,13 +92,7 @@ func _display_profile_list() -> void:
 
 func _load_profile(profile_name : String) -> void:
 	if delete_mode:
-		var dialog : MenuScreen = menu._add_screen("accept_dialog")
-		dialog.desc_text = "Are you sure you want to delete this profile?"
-		dialog.object_to_call = self
-		dialog.call_function_name = "_continue_deletion"
-		dialog.call_function_argument = profile_name
-		dialog.cancel_function_name = "_delete_profile"
-		return
+		_continue_deletion(profile_name)
 	
 	Data.profile._load_profile(profile_name)
 
@@ -113,10 +109,9 @@ func _load_profile(profile_name : String) -> void:
 		return
 
 	if Data.profile.status == Profile.STATUS.CONFIG_FAIL or Data.profile.status == Profile.STATUS.CONFIG_MISSING:
-		var dialog : MenuScreen = menu._add_screen("accept_dialog")
+		var dialog : MenuScreen = menu._add_screen("")
 		dialog.desc_text = "Warning! This profile config file is missing. Continue?"
-		dialog.object_to_call = self
-		dialog.call_function_name = "_continue_loading"
+		dialog.accept_function = _continue_loading
 		return
 	
 	menu.screens["foreground"].get_node("ProfileLayout/Name").text = profile_name
@@ -124,32 +119,36 @@ func _load_profile(profile_name : String) -> void:
 
 
 func _delete_profile() -> void:
-	if delete_mode:
+	delete_mode = true
+	$V/Text.text = "SELECT PROFILE TO DELETE"
+	$V/Text.modulate = Color.RED
+	$V/Menu/Create._disable(true)
+	$V/Menu/Delete.text = "CANCEL DELETION"
+
+
+func _continue_deletion(profile_name : String) -> void:
+	var dialog : MenuScreen = menu._add_screen("")
+	dialog.desc_text = "Are you sure you want to delete this profile?"
+	var accepted : bool = await dialog.closed
+
+	if accepted:
+		Data.profile._delete_profile(profile_name)
+		_delete_profile()
+		# Added delay so newly created buttons wont assign to accept menu which is going to be deleted
+		await get_tree().create_timer(0.8).timeout
+		_display_profile_list()
+		
+		if profile_name == Data.profile.name:
+			$V/Text.text = "YOU MUST SELECT OR CREATE AN PROFILE IN ORDER TO CONTINUE"
+			$V/Menu/Cancel._disable(true)
+	else:
 		delete_mode = false
 		$V/Text.text = "SELECT PROFILE TO LOAD"
 		$V/Text.modulate = Color.WHITE
 		$V/Menu/Create._disable(false)
 		$V/Menu/Delete.text = "DELETE PROFILE"
-	else:
-		if $V/Profiles/V.get_child_count() < 1 : return
-		
-		delete_mode = true
-		$V/Text.text = "SELECT PROFILE TO DELETE"
-		$V/Text.modulate = Color.RED
-		$V/Menu/Create._disable(true)
-		$V/Menu/Delete.text = "CANCEL DELETION"
 
-
-func _continue_deletion(profile_name : String) -> void:
-	Data.profile._delete_profile(profile_name)
-	_delete_profile()
-	# Added delay so newly created buttons wont assign to accept menu which is going to be deleted
-	await get_tree().create_timer(0.8).timeout
-	_display_profile_list()
-	
-	if profile_name == Data.profile.name:
-		$V/Text.text = "YOU MUST SELECT OR CREATE AN PROFILE IN ORDER TO CONTINUE"
-		$V/Menu/Cancel._disable(true)
+	return
 
 
 func _continue_loading() -> void:
@@ -160,9 +159,7 @@ func _continue_loading() -> void:
 func _start_profile_create() -> void:
 	var input : MenuScreen = menu._add_screen("text_input")
 	input.desc_text = "Enter new profile name"
-	input.object_to_call = self
-	input.call_function_name = "_create_profile"
-	input._start()
+	input.accept_function = _create_profile
 
 
 func _create_profile(profile_name : String) -> void:
